refactor(via): header value predicates are in mod value
